split-driver(04) [driver][NFC] Split conv module to conv 1/2/3 + common
